+++ launcher/AssertHelpers.h
